+++ projects/framework-common-extensions/release_notes.md
@@ -2,5 +2,6 @@
 
 ## Upcoming
 
+* [fixed] Disabled auto scrolling on asset list. Always select the first item on activation.
 * [changed] Added labels to publisher asset selector and publisher, aligned margins.
 * [new] Initial release.